+++ src/XMLlistAttr.cpp
@@ -99,10 +99,10 @@ XMLlistAttr(XMLlistAttrStruct *p){
 	//the error code
 	int err = 0;
 
-	extern std::map<int,igorXMLfile> allXMLfiles;
+	extern std::map<long,igorXMLfile> allXMLfiles;
 	xmlXPathObject *xpathObj = NULL; 
 	xmlDoc *doc = NULL;
-	int fileID = -1;
+	long fileID = -1;
 	
 	//the filename handle, Xpath handle,namespace handle,options handle
 	char *xPath = NULL;
@@ -134,7 +134,7 @@ XMLlistAttr(XMLlistAttrStruct *p){
 	if (err = GetCStringFromHandle(p->ns, ns, sizens))
 		goto done;
 	
-	fileID = (int)roundf(p->fileID);	
+	fileID = (long)roundf(p->fileID);	
 	if((allXMLfiles.find(fileID) == allXMLfiles.end())){
 		XOPNotice("XMLlistAttr: fileID isn't valid\r");
 		err = FILEID_DOESNT_EXIST;
