+++ src/XMLstrFmXPath.cpp
@@ -69,8 +69,8 @@ XMLstrFmXPath(XMLstrFmXpathStructPtr p){
 	//the output Handle
 	Handle output = NULL;
 	
-	extern std::map<int,igorXMLfile> allXMLfiles;
-	int fileID = -1;
+	extern std::map<long,igorXMLfile> allXMLfiles;
+	long fileID = -1;
 	xmlXPathObject *xpathObj = NULL; 
 	xmlDoc *doc = NULL;
 
@@ -118,7 +118,7 @@ XMLstrFmXPath(XMLstrFmXpathStructPtr p){
 		return NOMEM; 
     p->returnString = output;
 
-	fileID = (int)roundf(p->fileID);	
+	fileID = (long)roundf(p->fileID);	
 	if((allXMLfiles.find(fileID) == allXMLfiles.end())){
 		XOPNotice("XMLstrfmXPath: fileID isn't valid\r");
 		err = FILEID_DOESNT_EXIST;
